refactor: rename simple.py
