+++ autora/cycle/__init__.py
@@ -1,1 +1,1 @@
-from ._simple_cycle import SimpleCycle as Cycle
+from .simple import SimpleCycle as Cycle
